requirements.txt
numpy==1.17.2
Pillow==6.2.1
